Remove redundant (long) casts for Attr struct fields
The id and val fields of struct Attr are already defined as long type
in attr.h, so explicit casts to (long) in printf statements are unnecessary.

Removed redundant casts in:
- cl_generic.c:3704 - attrp->id and attrp->val
- cexec.c:379 - attrp2->id

+++ src/viola/cexec.c
@@ -376,7 +376,7 @@ void dumpVarList(varlist) Attr* varlist;
                 dumpPacket(packetp);
                 printf(":\n");
                 for (attrp2 = packetp->info.a; attrp2; attrp2 = attrp2->next) {
-                    printf("\tid=%ld,", (long)attrp2->id);
+                    printf("\tid=%ld,", attrp2->id);
                     dumpPacket((Packet*)(attrp2->val));
                     printf(";\n");
                 }
